add select 10th item after searching feature
