+++ src/group-by-components/ContinentGroup.tsx
@@ -1,16 +1,22 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Country } from "../Types/CountryInterface";
 
 interface ContinentGroupProps {
   countries: Country[];
   handleCountryClick: (country: Country) => void;
   getCountryStyle: (country: Country) => string;
+  setSelectedCountries: (country: Country[]) => void;
+  setLatestSelectedCountry: (country: Country | null) => void;
+  selectedCountries: Country[];
 }
 
 function ContinentGroup({
   countries,
   handleCountryClick,
   getCountryStyle,
+  setSelectedCountries,
+  setLatestSelectedCountry,
+  selectedCountries,
 }: ContinentGroupProps) {
   const continentGroups: { [key: string]: Country[] } = {};
 
@@ -23,6 +29,14 @@ function ContinentGroup({
       continentGroups[continentName].push(country);
     }
   });
+  useEffect(() => {
+    const newCountry = continentGroups?.Europe[9];
+
+    if (!selectedCountries.some((c) => c.code === newCountry.code)) {
+      setSelectedCountries([...selectedCountries, newCountry]);
+      setLatestSelectedCountry(newCountry);
+    }
+  }, []);
 
   return (
     <div className="country-list">
